Refactor tests to reflect on previous changes
